requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCL==1.3.0
discord.py[voice]
dnspython==1.16.0